pyproject.toml
[tool.poetry]
name = "data-science-admin-tool"
version = "0.1.0"
description = ""
authors = ["Mikey Lau <63963445+MikeyJL@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "data_science_admin_tool"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.3"
djangorestframework = "^3.14.0"
Markdown = "^3.4.1"
django-filter = "^22.1"
python-dotenv = "^0.21.0"
flake8-docstrings = "^1.6.0"
pycognito = "^2022.8.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
black = "^22.10.0"
pre-commit = "^2.20.0"
mypy = "^0.982"
django-stubs = "^1.12.0"
djangorestframework-stubs = "^1.7.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
